+++ luqum/elasticsearch/__init__.py
@@ -1,1 +1,2 @@
 from .visitor import ElasticsearchQueryBuilder  # noqa:
+from .schema import SchemaAnalyzer  # noqa: